+++ src/AsyncRepositories/IDeleteAsyncRepositoryBase.cs
@@ -2,13 +2,13 @@
 using System.Threading.Tasks;
 using PowerUtils.BuildingBlocks.Domain;
 
-namespace PowerUtils.BuildingBlocks.Data.Repositories
+namespace PowerUtils.BuildingBlocks.Data.AsyncRepositories
 {
-    public interface ICountAsyncRepositoryBase<TEntity, TId> :
+    public interface IDeleteAsyncRepositoryBase<TEntity, TId> :
         IEntityAsyncRepositoryBase<TEntity, TId>
 
         where TEntity : IAggregateRoot<TId>
     {
-        Task<long> CountAsync(CancellationToken cancellationToken = default);
+        Task DeleteAsync(TId id, CancellationToken cancellationToken = default);
     }
 }