+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Breaking Changes
 - Added "Async" word in interfaces for async methods;
+- Named the namespace for async method from `PowerUtils.BuildingBlocks.Data.Repositories` to `PowerUtils.BuildingBlocks.Data.AsyncRepositories`;